+++ src/main/kotlin/ru/testit/kotlin/client/models/WorkItemShortApiResult.kt
@@ -98,7 +98,7 @@ data class WorkItemShortApiResult (
 
     /* Work Item duration */
     @Json(name = "duration")
-    val duration: kotlin.Int,
+    val duration: kotlin.Long,
 
     /* Unique identifier of user who created Work Item */
     @Json(name = "createdById")
